refactor: move application entry point to root package com.occupi #58
