--- backend/src/main/java/com/occupi/app/AppApplication.java
@@ -1,15 +1,0 @@
-package com.occupi.app;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-
-@SpringBootApplication
-@ComponentScan(basePackages = "com.occupi")
-public class AppApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AppApplication.class, args);
-	}
-
-}
